fix: Help now uses alias name if it's set
Previously help would use the alias name from the core hubot but
external commands would use the full robot.name attribute. This now does
some logic to see if robot.alias is set and if so, use that vs the full
robot name. This is for consistency and sorting in the help output.

fixes #3